BE/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.235.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymssql==2.2.11
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.7
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pydantic==2.5.0
requests==2.31.0
python-keycloak==3.4.0
pydantic-settings==2.1.0
python-keycloak
aioodbc